// Proyecto_DBD/planillafrontend/src/interfaces.ts
export interface Empleado{
  id_empleado: number;
  nombres: string;
  apellidos: string;
  fecha_nacimiento: string;
  edad: number;
  telefono: number;
  email: string;
  direccion: string;
  dni: number;
  estado_civil: string;
  sexo: string;
  id_regimen_pensionario: number;
  id_seguro_medico: number;
  id_situacion_discapacidad: number;
  id_empresa: number;
  id_cuenta: number;
}